# public/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "有来管理后台 API",
        "description": "基于 ThinkPHP 8 的 RBAC 权限管理系统后端 API",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "开发环境"
        }
    ],
    "paths": {
...
        "/api/v1/auth/login": {
            "post": {
                "tags": [
                    "01.认证中心"
                ],
                "summary": "账号密码登录",
                "description": "登录",
                "operationId": "629e79825894b774f085d1b915dc1db9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/v1/auth/sms/code": {
            "post": {
                "tags": [
                    "01.认证中心"
                ],
                "summary": "发送登录短信验证码",
                "description": "发送登录短信验证码",
                "operationId": "136c205b814e2d46bdb75aee7bec0554",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "手机号",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
        "/api/v1/auth/login/sms": {
            "post": {
                "tags": [
                    "01.认证中心"
                ],
                "summary": "短信验证码登录",
                "description": "短信验证码登录",
                "operationId": "c9acfe519c7dab6b94e0e747c5f836f5",
                "parameters": [
                    {
                        "name": "mobile",
                        "in": "query",
                        "description": "手机号",
                        "required": true
                    },
                    {
                        "name": "code",
                        "in": "query",
                        "description": "验证码",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功"
                    }
                }
            }
        },
...
        "/api/v1/codegen/table": {
            "get": {
                "tags": [
                    "11.代码生成"
                ],
                "summary": "获取数据表分页列表",
                "description": "数据表分页",
                "operationId": "6863b7062848d2e086e9b9ad71f57123",
                "parameters": [
                    {
                        "name": "pageNum",
                        "in": "query",
                        "description": "页码",
                        "required": false
                    },
                    {
                        "name": "pageSize",
                        "in": "query",
                        "description": "每页数量",
                        "required": false
                    },
                    {
                        "name": "keywords",
                        "in": "query",
                        "description": "关键字",
                        "required": false
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/codegen/{tableName}/config": {
            "get": {
                "tags": [
                    "11.代码生成"
                ],
                "summary": "获取代码生成配置",
                "description": "获取生成配置",
                "operationId": "7cd8fb107947a3cc611f6f3d587f52e0",
                "parameters": [
                    {
                        "name": "tableName",
                        "in": "path",
                        "description": "表名",
                        "required": true,
                        "example": "sys_user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "post": {
                "tags": [
                    "11.代码生成"
                ],
                "summary": "保存代码生成配置",
                "description": "保存生成配置",
                "operationId": "b493c1ac8c2d3ecb41303bf74c3c7d64",
                "parameters": [
                    {
                        "name": "tableName",
                        "in": "path",
                        "description": "表名",
                        "required": true,
                        "example": "sys_user"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": []
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            },
            "delete": {
                "tags": [
                    "11.代码生成"
                ],
                "summary": "删除代码生成配置",
                "description": "删除生成配置",
                "operationId": "8830b47d604e8d8961d347068983dc89",
                "parameters": [
                    {
                        "name": "tableName",
                        "in": "path",
                        "description": "表名",
                        "required": true,
                        "example": "sys_user"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/codegen/{tableName}/preview": {
            "get": {
                "tags": [
                    "11.代码生成"
                ],
                "summary": "获取预览生成代码",
                "description": "预览生成代码",
                "operationId": "96ed8e328ef289bcd3c4c912a785fb3f",
                "parameters": [
                    {
                        "name": "tableName",
                        "in": "path",
                        "description": "表名",
                        "required": true,
                        "example": "sys_user"
                    },
                    {
                        "name": "pageType",
                        "in": "query",
                        "description": "页面类型",
                        "required": false,
                        "example": "classic"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "代码类型",
                        "required": false,
                        "example": "ts"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/api/v1/codegen/{tableName}/download": {
            "get": {
                "tags": [
                    "11.代码生成"
                ],
                "summary": "下载代码",
                "description": "下载 zip",
                "operationId": "376f9dd75606cc3840c44151e62c30c9",
                "parameters": [
                    {
                        "name": "tableName",
                        "in": "path",
                        "description": "表名(多个逗号分隔)",
                        "required": true,
                        "example": "sys_user"
                    },
                    {
                        "name": "pageType",
                        "in": "query",
                        "description": "页面类型",
                        "required": false,
                        "example": "classic"
                    },
                    {
                        "name": "type",
                        "in": "query",
                        "description": "代码类型",
                        "required": false,
                        "example": "ts"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
...
    },
...
}